add: changes in types of dosis and frecuenciaHoras, component added in agregarMedicamento

+++ mimedicapp/lib/models/medicamentoUsuario.dart
@@ -1,9 +1,11 @@
+import 'dart:ffi';
+
 class MedicamentoUsuario {
   final int? id;
   final String nombre;
-  final int dosis;
+  final double dosis;
   final String unidad;
-  final int frecuenciaHoras;
+  final double frecuenciaHoras;
   final DateTime fechaInicio;
   final DateTime fechaFin;
